+++ CHANGELOG.md
@@ -20,6 +20,10 @@ All notable changes to Botonic will be documented in this file.
 
   - [use new endpoint v2 to deploy bots](https://github.com/hubtype/botonic/pull/2884)
 
+- [@botonic/react](https://www.npmjs.com/package/@botonic/react)
+
+  - [remove getConfig function](https://github.com/hubtype/botonic/pull/2902)
+
 ### Fixed
 
 </details>